hospital-api/requirements.txt
fastapi>=0.104.1
uvicorn==0.24.0
mlflow==2.8.0
pandas>=2.1.0,<3
numpy==1.24.4
python-jose==3.3.0
python-dotenv==1.0.0
httpx==0.25.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
prophet
xgboost
databricks-sdk
shap
